+++ cepc_vue/src/views/Records.vue
@@ -1,56 +1,61 @@
 <template>
   <div>
-    <el-table
-        :data="tableData"
-        border
-        style="width: 100%">
-      <el-table-column
-          fixed
-          prop="id"
-          label="编号"
-          width="60">
-      </el-table-column>
-      <el-table-column
-          prop="name"
-          label="姓名"
-          width="120">
-      </el-table-column>
-      <el-table-column
-          prop="temperature"
-          label="体温"
-          width="80">
-      </el-table-column>
-      <el-table-column
-          prop="patient"
-          label="何类人员"
-          width="120">
-      </el-table-column>
-      <el-table-column
-          prop="date"
-          label="日期"
-          width="120">
-      </el-table-column>
-      <el-table-column
-          prop="address"
-          label="地址">
-      </el-table-column>
-      <el-table-column
-          fixed="right"
-          label="操作"
-          width="130">
-        <template slot-scope="scope">
-          <el-button @click="edit(scope.row)" type="text" size="small">修改</el-button>
-          <el-button @click="deleteRecord(scope.row)" type="text" size="small">删除</el-button>
-        </template>
-      </el-table-column>
-    </el-table>
-    <el-pagination
-        background
-        layout="prev, pager, next"
-        :page-size="totalSize"
-        :total="totalElement"
-        @current-change="page">
-    </el-pagination>
+    <el-col style="width: 90%">
+      <el-table
+          :data="tableData"
+          border
+          style="width: 100%">
+        <el-table-column
+            fixed
+            prop="id"
+            label="编号"
+            width="60">
+        </el-table-column>
+        <el-table-column
+            prop="name"
+            label="姓名"
+            width="120">
+        </el-table-column>
+        <el-table-column
+            prop="temperature"
+            label="体温"
+            width="80">
+        </el-table-column>
+        <el-table-column
+            prop="patient"
+            label="何类人员"
+            width="100">
+        </el-table-column>
+        <el-table-column
+            prop="date"
+            label="日期"
+            width="120">
+        </el-table-column>
+        <el-table-column
+            prop="address"
+            label="地址">
+        </el-table-column>
+        <el-table-column
+            fixed="right"
+            label="操作"
+            width="130">
+          <template slot-scope="scope">
+            <el-button @click="edit(scope.row)" type="primary" icon="el-icon-edit" circle size="small"></el-button>
+            <el-button @click="deleteRecord(scope.row)" type="danger" icon="el-icon-delete" circle size="small"></el-button>
+          </template>
+        </el-table-column>
+      </el-table>
+      <el-pagination
+          background
+          layout="prev, pager, next"
+          :page-size="totalSize"
+          :total="totalElement"
+          @current-change="page">
+      </el-pagination>
+    </el-col>
+    <el-col style="width: 10%">
+      <el-button @click="add()" icon="el-icon-plus" circle ></el-button>
+    </el-col>
   </div>
 </template>
 
@@ -59,13 +64,29 @@ export default {
   methods: {
     deleteRecord(row){
       const _this = this
-      axios.delete('http://localhost:8021/records/deleteById/'+row.id).then(function (resp) {
-        _this.$alert(row.name+'删除成功', '消息', {
-          confirmButtonText: '确定',
-          callback: action => {
-            window.location.reload()
-          }
+      _this.$confirm('此操作将永久删除该数据, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        axios.delete('http://localhost:8021/records/deleteById/'+row.id).then(function (resp) {
+          _this.$message({
+            type: 'success',
+            message: '删除成功!'
+          });
+        })
+        window.location.reload()
+      }).catch(() => {
+        this.$message({
+          type: 'info',
+          message: '已取消删除'
         });
+      });
+    },
+    add(){
+      const _this = this
+      _this.$router.push({
+        path:'/addRecord'
       })
     },
     edit(row) {
